Refactored encounter class into 2 classes
